added video mode support, adde new ui controls

+++ core/time.h
@@ -2,4 +2,4 @@
 * @brief Returns current process time in seconds
 * @return time in seconds
 */
-float de_time_get_seconds(void);
+double de_time_get_seconds(void);
