+++ physics/impl/collision.h
@@ -19,9 +19,7 @@
 * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE. */
 
-/*=======================================================================================*/
-void de_static_geometry_add_triangle(de_static_geometry_t* geom, const de_vec3_t* a, const de_vec3_t* b, const de_vec3_t* c)
-{
+void de_static_geometry_add_triangle(de_static_geometry_t* geom, const de_vec3_t* a, const de_vec3_t* b, const de_vec3_t* c) {
 	de_vec3_t ca;
 	de_static_triangle_t triangle;
 
@@ -55,24 +53,19 @@ void de_static_geometry_add_triangle(de_static_geometry_t* geom, const de_vec3_t
 	DE_ARRAY_APPEND(geom->triangles, triangle);
 }
 
-/*=======================================================================================*/
-void de_static_geometry_fill(de_static_geometry_t* geom, const de_mesh_t* mesh, const de_mat4_t transform)
-{
+void de_static_geometry_fill(de_static_geometry_t* geom, const de_mesh_t* mesh, const de_mat4_t transform) {
 	size_t i, k;
 
-	if (!geom || !mesh)
-	{
+	if (!geom || !mesh) {
 		return;
 	}
 
-	for (i = 0; i < mesh->surfaces.size; ++i)
-	{
+	for (i = 0; i < mesh->surfaces.size; ++i) {
 		de_surface_t* surf;
 
 		surf = mesh->surfaces.data[i];
 
-		for (k = 0; k < surf->indices.size; k += 3)
-		{
+		for (k = 0; k < surf->indices.size; k += 3) {
 			de_vertex_t va, vb, vc;
 			de_vec3_t pa, pb, pc;
 			int a, b, c;
@@ -98,9 +91,7 @@ void de_static_geometry_fill(de_static_geometry_t* geom, const de_mesh_t* mesh, 
 	}
 }
 
-/*=======================================================================================*/
-static bool de_static_triangle_contains_point(const de_static_triangle_t* triangle, const de_vec3_t* point)
-{
+static bool de_static_triangle_contains_point(const de_static_triangle_t* triangle, const de_vec3_t* point) {
 	de_vec3_t vp;
 	float dot02, dot12, u, v;
 
@@ -112,55 +103,45 @@ static bool de_static_triangle_contains_point(const de_static_triangle_t* triang
 	return (u >= 0.0f) && (v >= 0.0f) && (u + v < 1.0f);
 }
 
-/*=======================================================================================*/
-static bool de_is_point_on_line_segment(const de_vec3_t* point, const de_vec3_t* a, const de_vec3_t* b)
-{
+static bool de_is_point_on_line_segment(const de_vec3_t* point, const de_vec3_t* a, const de_vec3_t* b) {
 	/* simply check, if point lies in bounding box (a,b), means that point is on line  */
 	de_vec3_t min = *a;
 	de_vec3_t max = *b;
 
 	/* swap coordinates if needed */
 	float temp;
-	if (min.x > max.x)
-	{
+	if (min.x > max.x) {
 		temp = min.x;
 		min.x = max.x;
 		max.x = temp;
 	}
-	if (min.y > max.y)
-	{
+	if (min.y > max.y) {
 		temp = min.y;
 		min.y = max.y;
 		max.y = temp;
 	}
-	if (min.z > max.z)
-	{
+	if (min.z > max.z) {
 		temp = min.z;
 		min.z = max.z;
 		max.z = temp;
 	}
 
-	if ((point->x > max.x) || (point->y > max.y) || (point->z > max.z))
-	{
+	if ((point->x > max.x) || (point->y > max.y) || (point->z > max.z)) {
 		return false;
 	}
-	if ((point->x < min.x) || (point->y < min.y) || (point->z < min.z))
-	{
+	if ((point->x < min.x) || (point->y < min.y) || (point->z < min.z)) {
 		return false;
 	}
 	return true;
 }
 
-/*=======================================================================================*/
-static void de_project_point_on_line(const de_vec3_t* point, const de_ray_t* ray, de_vec3_t* out)
-{
+static void de_project_point_on_line(const de_vec3_t* point, const de_ray_t* ray, de_vec3_t* out) {
 	float sqr_len;
 	de_vec3_t pa, offset;
 
 	sqr_len = de_vec3_sqr_len(&ray->dir);
 
-	if (sqr_len < 0.0001f)
-	{
+	if (sqr_len < 0.0001f) {
 		de_vec3_zero(out);
 	}
 
@@ -169,44 +150,34 @@ static void de_project_point_on_line(const de_vec3_t* point, const de_ray_t* ray
 	de_vec3_add(out, &ray->origin, &offset);
 }
 
-/*=======================================================================================*/
-static bool de_body_sphere_intersection(const de_ray_t* edgeRay, const de_vec3_t* sphere_pos, float sphere_radius, de_vec3_t* intersection_pt)
-{
+static bool de_body_sphere_intersection(const de_ray_t* edgeRay, const de_vec3_t* sphere_pos, float sphere_radius, de_vec3_t* intersection_pt) {
 	de_vec3_t ray_end_pt;
-	if (de_ray_sphere_intersection(edgeRay, sphere_pos, sphere_radius, NULL, NULL))
-	{
+	if (de_ray_sphere_intersection(edgeRay, sphere_pos, sphere_radius, NULL, NULL)) {
 		de_project_point_on_line(sphere_pos, edgeRay, intersection_pt);
 		de_vec3_add(&ray_end_pt, &edgeRay->origin, &edgeRay->dir);
-		if (de_is_point_on_line_segment(intersection_pt, &edgeRay->origin, &ray_end_pt))
-		{
+		if (de_is_point_on_line_segment(intersection_pt, &edgeRay->origin, &ray_end_pt)) {
 			return true;
 		}
 	}
 	return false;
 }
 
-/*=======================================================================================*/
-static bool de_body_point_intersection(const de_vec3_t* point, const de_vec3_t* sphere_pos, float sphere_radius, de_vec3_t* intersection_pt)
-{
-	if (de_vec3_sqr_distance(point, sphere_pos) < sphere_radius * sphere_radius)
-	{
+static bool de_body_point_intersection(const de_vec3_t* point, const de_vec3_t* sphere_pos, float sphere_radius, de_vec3_t* intersection_pt) {
+	if (de_vec3_sqr_distance(point, sphere_pos) < sphere_radius * sphere_radius) {
 		*intersection_pt = *point;
 		return true;
 	}
 	return false;
 }
 
-/*=======================================================================================*/
-static bool de_sphere_triangle_intersection(const de_vec3_t* sphere_pos, float sphere_radius, const de_static_triangle_t* triangle, de_vec3_t* intersection_pt)
-{
+static bool de_sphere_triangle_intersection(const de_vec3_t* sphere_pos, float sphere_radius, const de_static_triangle_t* triangle, de_vec3_t* intersection_pt) {
 	de_plane_t plane;
 	float distance;
 
 	de_plane_set(&plane, &triangle->normal, triangle->distance);
 
 	distance = de_plane_distance(&plane, sphere_pos);
-	if (distance <= sphere_radius)
-	{
+	if (distance <= sphere_radius) {
 		de_vec3_t offset;
 		de_vec3_scale(&offset, &plane.n, distance);
 		de_vec3_sub(intersection_pt, sphere_pos, &offset);
@@ -221,20 +192,15 @@ static bool de_sphere_triangle_intersection(const de_vec3_t* sphere_pos, float s
 	return false;
 }
 
-/*=======================================================================================*/
-static void de_body_verlet(de_body_t* body, float dt2)
-{
+static void de_body_verlet(de_body_t* body, float dt2) {
 	de_vec3_t last_position;
 	float friction, k1, k2;
 	de_vec3_t velocity;
 	float velocity_limit = 0.75f;
 
-	if (body->contact_count > 0)
-	{
+	if (body->contact_count > 0) {
 		friction = 1 - body->friction;
-	}
-	else
-	{
+	} else {
 		friction = DE_AIR_FRICTION;
 	}
 
@@ -255,36 +221,27 @@ static void de_body_verlet(de_body_t* body, float dt2)
 
 	/* Velocity limiting */
 	de_vec3_sub(&velocity, &body->last_position, &body->position);
-
-	if (de_vec3_sqr_len(&velocity) > velocity_limit * velocity_limit)
-	{
+	if (de_vec3_sqr_len(&velocity) > velocity_limit * velocity_limit) {
 		de_vec3_normalize(&velocity, &velocity);
 		de_vec3_scale(&velocity, &velocity, velocity_limit);
-
 		de_vec3_sub(&body->last_position, &body->position, &velocity);
 	}
 }
 
-/*=======================================================================================*/
-static de_contact_t* de_body_add_contact(de_body_t* body)
-{
-	if (body->contact_count < DE_MAX_CONTACTS)
-	{
+static de_contact_t* de_body_add_contact(de_body_t* body) {
+	if (body->contact_count < DE_MAX_CONTACTS) {
 		return &body->contacts[body->contact_count++];
 	}
 	return NULL;
 }
 
-/*=======================================================================================*/
-void de_body_triangle_collision(de_static_triangle_t* triangle, de_body_t* sphere)
-{
+void de_body_triangle_collision(de_static_triangle_t* triangle, de_body_t* sphere) {
 	de_vec3_t intersectionPoint;
 	de_vec3_t middle, orientation, offset;
 	float length, penetrationDepth;
 	de_contact_t* contact;
 
-	if (de_sphere_triangle_intersection(&sphere->position, sphere->radius, triangle, &intersectionPoint))
-	{
+	if (de_sphere_triangle_intersection(&sphere->position, sphere->radius, triangle, &intersectionPoint)) {
 		length = 0.0f;
 
 		/* Calculate penetration depth and push vector */
@@ -293,8 +250,7 @@ void de_body_triangle_collision(de_static_triangle_t* triangle, de_body_t* spher
 		penetrationDepth = sphere->radius - length;
 
 		/* Degenerated case, ignore */
-		if (penetrationDepth < 0.0f)
-		{
+		if (penetrationDepth < 0.0f) {
 			return;
 		}
 
@@ -304,8 +260,7 @@ void de_body_triangle_collision(de_static_triangle_t* triangle, de_body_t* spher
 		/* Write contact info */
 		contact = de_body_add_contact(sphere);
 
-		if (contact)
-		{
+		if (contact) {
 			contact->body = NULL;
 			contact->normal = orientation;
 			contact->position = intersectionPoint;
@@ -314,36 +269,24 @@ void de_body_triangle_collision(de_static_triangle_t* triangle, de_body_t* spher
 	}
 }
 
-/*=======================================================================================*/
-void de_physics_step(de_core_t* core, float dt)
-{
+void de_physics_step(de_core_t* core, double dt) {
 	float dt2;
 	de_scene_t* scene;
 	de_body_t* body;
 	de_static_geometry_t* geom;
-
-	dt2 = dt * dt;
-
-	DE_LINKED_LIST_FOR_EACH(core->scenes, scene)
-	{
-		DE_LINKED_LIST_FOR_EACH(scene->bodies, body)
-		{
+	dt2 = (float)(dt * dt);
+	DE_LINKED_LIST_FOR_EACH(core->scenes, scene) {
+		DE_LINKED_LIST_FOR_EACH(scene->bodies, body) {
 			/* Drop contact information */
 			body->contact_count = 0;
-
 			/* Apply gravity */
 			de_vec3_add(&body->acceleration, &body->acceleration, &body->gravity);
-
 			/* Do Verlet integration */
 			de_body_verlet(body, dt2);
-
 			/* Solve collisions */
-			DE_LINKED_LIST_FOR_EACH(scene->static_geometries, geom)
-			{
+			DE_LINKED_LIST_FOR_EACH(scene->static_geometries, geom) {
 				size_t j;
-
-				for (j = 0; j < geom->triangles.size; ++j)
-				{
+				for (j = 0; j < geom->triangles.size; ++j) {
 					de_body_triangle_collision(&geom->triangles.data[j], body);
 				}
 			}
@@ -351,64 +294,45 @@ void de_physics_step(de_core_t* core, float dt)
 	}
 }
 
-/*=======================================================================================*/
-void de_body_set_gravity(de_body_t * body, const de_vec3_t * gravity)
-{
+void de_body_set_gravity(de_body_t * body, const de_vec3_t * gravity) {
 	body->gravity = *gravity;
 }
 
-/*=======================================================================================*/
-void de_body_set_position(de_body_t * body, const de_vec3_t * pos)
-{
+void de_body_set_position(de_body_t * body, const de_vec3_t * pos) {
 	body->position = *pos;
 	body->last_position = *pos;
 }
 
-/*=======================================================================================*/
-void de_body_set_radius(de_body_t * body, float radius)
-{
+void de_body_set_radius(de_body_t * body, float radius) {
 	body->radius = radius;
 }
 
-/*=======================================================================================*/
-float de_body_get_radius(de_body_t* body)
-{
+float de_body_get_radius(de_body_t* body) {
 	return body->radius;
 }
 
-/*=======================================================================================*/
-size_t de_body_get_contact_count(de_body_t* body)
-{
+size_t de_body_get_contact_count(de_body_t* body) {
 	return body->contact_count;
 }
 
-/*=======================================================================================*/
-de_contact_t* de_body_get_contact(de_body_t* body, size_t i)
-{
-	if (i >= DE_MAX_CONTACTS)
-	{
+de_contact_t* de_body_get_contact(de_body_t* body, size_t i) {
+	if (i >= DE_MAX_CONTACTS) {
 		return NULL;
 	}
 
 	return body->contacts + i;
 }
 
-/*=======================================================================================*/
-void de_body_move(de_body_t * body, const de_vec3_t* velocity)
-{
-	if (!body || !velocity)
-	{
+void de_body_move(de_body_t * body, const de_vec3_t* velocity) {
+	if (!body || !velocity) {
 		return;
 	}
 
 	de_vec3_add(&body->position, &body->position, velocity);
 }
 
-/*=======================================================================================*/
-void de_body_set_velocity(de_body_t* body, const de_vec3_t* velocity)
-{
-	if (!body || !velocity)
-	{
+void de_body_set_velocity(de_body_t* body, const de_vec3_t* velocity) {
+	if (!body || !velocity) {
 		return;
 	}
 
@@ -417,44 +341,32 @@ void de_body_set_velocity(de_body_t* body, const de_vec3_t* velocity)
 	de_vec3_sub(&body->last_position, &body->last_position, velocity);
 }
 
-/*=======================================================================================*/
-void de_body_set_y_velocity(de_body_t * body, float y_velocity)
-{
-	if (!body)
-	{
+void de_body_set_y_velocity(de_body_t * body, float y_velocity) {
+	if (!body) {
 		return;
 	}
 
 	body->last_position.y = body->position.y - y_velocity;
 }
 
-/*=======================================================================================*/
-void de_body_set_x_velocity(de_body_t * body, float x_velocity)
-{
-	if (!body)
-	{
+void de_body_set_x_velocity(de_body_t * body, float x_velocity) {
+	if (!body) {
 		return;
 	}
 
 	body->last_position.x = body->position.x - x_velocity;
 }
 
-/*=======================================================================================*/
-void de_body_set_z_velocity(de_body_t * body, float z_velocity)
-{
-	if (!body)
-	{
+void de_body_set_z_velocity(de_body_t * body, float z_velocity) {
+	if (!body) {
 		return;
 	}
 
 	body->last_position.z = body->position.z - z_velocity;
 }
 
-/*=======================================================================================*/
-void de_body_get_velocity(de_body_t * body, de_vec3_t * velocity)
-{
-	if (!body || velocity)
-	{
+void de_body_get_velocity(de_body_t * body, de_vec3_t * velocity) {
+	if (!body || velocity) {
 		return;
 	}
 	de_vec3_sub(velocity, &body->position, &body->last_position);
